+++ examples/blazor-faktur/Program.cs
@@ -1,0 +1,7 @@
+using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
+using FoundryBlazorFaktur.Pages;
+
+var builder = WebAssemblyHostBuilder.CreateDefault(args);
+builder.RootComponents.Add<Faktur>("#app");
+
+await builder.Build().RunAsync();